Refactor: Enforce AGENTS.md structure and clean up artifacts
- Moved shared domain models to src/coreason_manifest/spec/common/.
- Moved V2 models to src/coreason_manifest/spec/v2/.
- Moved V2 logic to src/coreason_manifest/utils/v2/.
- Moved root models to src/coreason_manifest/spec/.
- Updated all imports in source, tests, scripts, and documentation.
- Deleted obsolete scripts and artifacts.
- Verified 100% test coverage and strict typing.
