--- requirements.txt
@@ -1,14 +1,0 @@
-annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "3.15"
-attrs==25.4.0 ; python_version >= "3.12" and python_version < "3.15"
-colorama==0.4.6 ; python_version >= "3.12" and python_version < "3.15" and sys_platform == "win32"
-jsonschema-specifications==2025.9.1 ; python_version >= "3.12" and python_version < "3.15"
-jsonschema==4.26.0 ; python_version >= "3.12" and python_version < "3.15"
-loguru==0.7.3 ; python_version >= "3.12" and python_version < "3.15"
-pydantic-core==2.41.5 ; python_version >= "3.12" and python_version < "3.15"
-pydantic==2.12.5 ; python_version >= "3.12" and python_version < "3.15"
-pyyaml==6.0.3 ; python_version >= "3.12" and python_version < "3.15"
-referencing==0.37.0 ; python_version >= "3.12" and python_version < "3.15"
-rpds-py==0.30.0 ; python_version >= "3.12" and python_version < "3.15"
-typing-extensions==4.15.0 ; python_version >= "3.12" and python_version < "3.15"
-typing-inspection==0.4.2 ; python_version >= "3.12" and python_version < "3.15"
-win32-setctime==1.2.0 ; python_version >= "3.12" and python_version < "3.15" and sys_platform == "win32"
