+++ scripts/validate_v2_sample.py
@@ -6,7 +6,7 @@ import yaml
 # Ensure src is in python path
 sys.path.append(str(Path(__file__).parent.parent / "src"))
 
-from coreason_manifest.v2.spec.definitions import ManifestV2
+from coreason_manifest.spec.v2.definitions import ManifestV2
 
 
 def validate_sample() -> None:
