+++ src/coreason_manifest/spec/common/error.py
