+++ src/coreason_manifest/spec/common/identity.py
@@ -12,7 +12,7 @@ from typing import Optional
 
 from pydantic import ConfigDict, Field
 
-from ..common import CoReasonBaseModel
+from ..common_base import CoReasonBaseModel
 
 
 class Identity(CoReasonBaseModel):
