+++ src/coreason_manifest/spec/common_base.py
