--- src/coreason_manifest/v2/__init__.py
@@ -1,1 +1,0 @@
-# Coreason Manifest V2
