--- test_walk/link
@@ -1,1 +1,0 @@
-/app/test_walk/target
