+++ tests/test_docs_examples.py
@@ -1,10 +1,10 @@
 from coreason_manifest import Manifest
-from coreason_manifest.v2.spec.contracts import (
+from coreason_manifest.spec.v2.contracts import (
     InterfaceDefinition,
     PolicyDefinition,
     StateDefinition,
 )
-from coreason_manifest.v2.spec.definitions import (
+from coreason_manifest.spec.v2.definitions import (
     AgentDefinition,
     AgentStep,
     ManifestMetadata,
