+++ tests/test_v2_agent_complex.py
@@ -1,6 +1,6 @@
 import yaml
 
-from coreason_manifest.v2.spec.definitions import (
+from coreason_manifest.spec.v2.definitions import (
     AgentDefinition,
     GenericDefinition,
     ManifestV2,
