+++ tests/test_v2_edge_cases.py
@@ -1,7 +1,7 @@
 import pytest
 
-from coreason_manifest.v2.spec.definitions import ManifestV2
-from coreason_manifest.v2.validator import validate_integrity
+from coreason_manifest.spec.v2.definitions import ManifestV2
+from coreason_manifest.utils.v2.validator import validate_integrity
 
 
 def test_missing_start_step() -> None:
