+++ tests/test_v2_phase3.py
@@ -10,10 +10,9 @@
 
 import pytest
 
-from coreason_manifest.common import ToolRiskLevel
-from coreason_manifest.governance import GovernanceConfig
-from coreason_manifest.v2.governance import check_compliance_v2
-from coreason_manifest.v2.spec.definitions import (
+from coreason_manifest.spec.common_base import ToolRiskLevel
+from coreason_manifest.spec.governance import GovernanceConfig
+from coreason_manifest.spec.v2.definitions import (
     AgentDefinition,
     AgentStep,
     ManifestMetadata,
@@ -21,7 +20,8 @@ from coreason_manifest.v2.spec.definitions import (
     ToolDefinition,
     Workflow,
 )
-from coreason_manifest.v2.validator import validate_integrity
+from coreason_manifest.utils.v2.governance import check_compliance_v2
+from coreason_manifest.utils.v2.validator import validate_integrity
 
 
 @pytest.fixture
